refactor: add avs service wrapper ABI instead of manually defining an interface and rename abi files
