+++ bindings/src/avs_service_manager_wrapper.rs
@@ -1,0 +1,6 @@
+// AvsServiceManagerWrapper contract bindings generated at compile time from ABI
+alloy::sol! {
+    #[sol(rpc)]
+    AvsServiceManagerWrapper,
+    "abis/AvsServiceManagerWrapper.abi.json"
+}
