+++ bindings/src/bls_sig_check_operator_state_retriever.rs
