--- bindings/src/interfaces/mod.rs
@@ -1,1 +1,0 @@
-pub mod avs_threshold;
